requirements.txt
pyjwt
graphlit-client
streamlit_extras
plotly>=5.0.0
streamlit_option_menu==0.3.13
pandas==2.2.2
requests==2.31.0
snowflake-connector-python==3.12.1
streamlit==1.38.0
cryptography==42.0.5
python-dotenv==1.0.1
snowflake-snowpark-python==1.21.1
streamlit-ace==0.1.1
